Find a Long Term Ecological Research (LTER) Site-Derived Palette

From a dataframe of all possible palettes (updated periodically so check back!) specify the characteristics of the palette you want and retrieve the palettes that match those criteria. Can specify by number of colors in the palette, type of palette (e.g., qualitative, sequential, etc.), or which LTER site the palette came from.

Usage

palette_find(site = "all", name = "all", type = "all")

Arguments

site

(character) Vector of three-letter LTER site abbreviations for which to return palettes or "all" or "LTER" for the LTER logo colors

name

(character) Vector of palette names (if known) for which to return palettes

type

(character) Vector of palette types (i.e., qualitative, tricolor, sequential, or diverging) for which to return palettes

Value

(dataframe / character) If more than one palette, a dataframe is returned; if exactly one palette, a character vector is returned

Examples

# Look at all palette options by calling the function without specifying arguments
lterpalettefinder::palette_find()
#> Multiple options returned as a dataframe. Consider identifying a specific palette's name and re-running function
#>           photographer    palette_full_name lter_site     palette_name
#> 1      Lina DiGregorio              AND-reu       AND              reu
#> 2        Chris Cousins       AND-salamander       AND       salamander
#> 3      Joelle Worthley         AND-training       AND         training
#> 4         Jansen Nipko           ARC-aurora       ARC           aurora
#> 5    Lindsay VanFossen           ARC-autumn       ARC           autumn
#> 6         Jansen Nipko           ARC-bucket       ARC           bucket
#> 7          Abigail Rec             ARC-hike       ARC             hike
#> 8    Lindsay VanFossen          ARC-sunrise       ARC          sunrise
#> 9    Lindsay VanFossen           ARC-sunset       ARC           sunset
#> 10       Alex Phillips       ASM-2022 group       ASM       2022 group
#> 11          Ken Dunton        BLE-ice drill       BLE        ice drill
#> 12        Kaylie Plumb           BLE-sunset       BLE           sunset
#> 13      Quincy Stewart           CAP-cactus       CAP           cactus
#> 14        Maddy Gibson         CAP-lovebird       CAP         lovebird
#> 15         Ashley Lang           HBR-canopy       HBR           canopy
#> 16         Ashley Lang            HBR-lotus       HBR            lotus
#> 17         Ashley Lang   HBR-mushroom floor       HBR   mushroom floor
#> 18      Jackie Matthes    HBR-mushroom tree       HBR    mushroom tree
#> 19      Lindsey Rustad      HBR-winter tree       HBR      winter tree
#> 20             Tatiana             HFR-newt       HFR             newt
#> 21 Augustín León-Sáenz           HFR-sunset       HFR           sunset
#> 22        Dylan Stover             JRN-calf       JRN             calf
#> 23      Ryan Schroeder          JRN-monsoon       JRN          monsoon
#> 24      Ryan Schroeder             JRN-oryx       JRN             oryx
#> 25 Brandon Bestelmeyer        JRN-pronghorn       JRN        pronghorn
#> 26      Ryan Schroeder          JRN-rainbow       JRN          rainbow
#> 27      Ryan Schroeder           JRN-sunset       JRN           sunset
#> 28       Michaela Rose            KBS-birds       KBS            birds
#> 29      G.P. Robertson             KBS-burn       KBS             burn
#> 30        Yahn-Jauh Su             KBS-fire       KBS             fire
#> 31         Kara Dobson          KBS-morning       KBS          morning
#> 32        Yahn-Jauh Su            KBS-storm       KBS            storm
#> 33     Corinn Rutkoski      KBS-swallowtail       KBS      swallowtail
#> 34         Jill Haukos            KNZ-bison       KNZ            bison
#> 35         Jill Haukos             KNZ-burn       KNZ             burn
#> 36         Jill Haukos       KNZ-night fire       KNZ       night fire
#> 37         Jill Haukos            KNZ-regal       KNZ            regal
#> 38 LTER Network Office            LTER-logo      LTER             logo
#> 39     Kathryn Scafidi           MCR-lagoon       MCR           lagoon
#> 40       Emily Stidham             NGA-boat       NGA             boat
#> 41         Carl Bowser            NTL-lakes       NTL            lakes
#> 42     Chiara Forester    NWT-mountain dawn       NWT    mountain dawn
#> 43     Chiara Forester   NWT-mountain field       NWT   mountain field
#> 44     Chiara Forester NWT-mountain flowers       NWT mountain flowers
#> 45     Chiara Forester   NWT-snowy mountain       NWT   snowy mountain
#> 46     Chiara Forester      NWT-snowy trees       NWT      snowy trees
#> 47     Chiara Forester      NWT-wildflowers       NWT      wildflowers
#> 48        Andrew Corso             PAL-boat       PAL             boat
#> 49        Andrew Corso              PAL-net       PAL              net
#> 50        Andrew Corso         PAL-penguins       PAL         penguins
#> 51        Andrew Corso            PAL-squid       PAL            squid
#> 52            SBC LTER      SBC-kelp forest       SBC      kelp forest
#> 53          Kyle Emery         SBC-kelp tag       SBC         kelp tag
#> 54       Amelie Berger             VCR-boat       VCR             boat
#> 55      Sophia Hoffman            VCR-marsh       VCR            marsh
#>    palette_type color_1 color_2 color_3 color_4 color_5 color_6 color_7 color_8
#> 1   qualitative #de696a #535150 #87715d #9e8873 #b0a58e #f1f0ed #959491 #69644d
#> 2      tricolor #6c616a #957b8e #bfa3b7 #ab8da1 #8a686f #b17b82 #7b5350 #af6a5f
#> 3    sequential #e37f64 #c2b9b0 #ded4c7 #e6c392 #959088 #b2a88f #938867 #6e674c
#> 4    sequential #a9ef66 #90e051 #c4fa9b #72cc4b #abf58c #98f082 #59af49 #87e77b
#> 5      tricolor #675a67 #c1bcbe #574a4f #7c7175 #918785 #d6d0ce #72564b #7e685d
#> 6   qualitative #6c696c #7d787b #918a8d #a49da0 #e8dee0 #4e4c4c #cac2c2 #965e4a
#> 7     diverging #5f5d5e #b6b3b2 #60514a #726860 #898480 #84766a #745e4b #d0c8bf
#> 8     diverging #b994b3 #a47396 #b9809c #d89eae #d08e9c #dccfd1 #987372 #cc7b76
#> 9    sequential #b04b44 #c14e44 #d45344 #e65945 #f75c44 #fc6a48 #e96d4a #fb7e54
#> 10     tricolor #c67157 #da936e #ebab86 #94694f #e6c3a0 #71604f #a3835f #b79875
#> 11  qualitative #da5a57 #76544d #90665a #be9587 #a97b68 #e2c5af #f1d060 #b8d5f1
#> 12    diverging #605d5d #9d9896 #aea9a6 #837973 #b98158 #dd9057 #916e54 #746c66
#> 13    diverging #e35769 #bf5159 #f48082 #936a4b #d1b6a1 #ae784b #5e5248 #b7875e
#> 14  qualitative #504f50 #7c7777 #dc6d53 #af7a63 #c0957a #8a705e #cbb9ac #9e8b7c
#> 15   sequential #f4f1ec #e4e2d9 #bab9b0 #a3a39d #cecec5 #e3e6ba #dfe790 #dcea6c
#> 16  qualitative #d6a7b7 #e5d0d7 #cc92a3 #e1bbc5 #c78189 #f0e6e4 #dd795a #f7e0cb
#> 17     tricolor #8b8186 #75686a #f98259 #5f534d #ab9284 #c0aca0 #856450 #eba67a
#> 18     tricolor #996f4f #be8254 #e09458 #ecb57c #75634c #f2dcb5 #b7a585 #d0c3a0
#> 19    diverging #645d61 #504c4e #9c9196 #796a6d #957776 #6d4f4e #896057 #6c7e8c
#> 20     tricolor #564d4f #765b4c #9a7052 #b98254 #b39e8c #806f5f #ac8e71 #908172
#> 21     tricolor #f4c978 #d9ba7d #e1cca2 #caa55a #a48e55 #fae090 #c5bb9a #837a4c
#> 22  qualitative #c1bbc1 #c4575a #f0eeee #95534c #8f8783 #aca19b #735a4b #806857
#> 23    diverging #bf9376 #ab8264 #987252 #826449 #bea98f #a0967b #8e8768 #7a7861
#> 24   sequential #c4885b #ae794f #d99a68 #ccbaab #976d4a #ecad77 #b4ada7 #c3a990
#> 25   sequential #87694b #cdb292 #957c5b #9f8c6a #ac9c7a #696148 #b4ab8a #565348
#> 26    diverging #666366 #524f51 #b9a49a #807369 #d1b6a0 #98826d #af9579 #6b5b48
#> 27    diverging #756575 #897b86 #7a5255 #986d64 #b37453 #ae8d77 #cb8a52 #eccaaa
#> 28   sequential #a6b8cf #b3c4da #8c9aad #7b95ba #99adca #9ba7b9 #8aa0c1 #6c87b0
#> 29  qualitative #b1a9b0 #544f52 #ad9092 #d39796 #f0b1a9 #785550 #a3736c #c37f74
#> 30     tricolor #ef9074 #cb7454 #f27e50 #a7664c #554e4a #705e4d #866c4e #9a7e5c
#> 31    diverging #c1b2a5 #d7c8b7 #b69e81 #d6b285 #f5c076 #ebae53 #c99c54 #f1d19a
#> 32    diverging #b39b58 #c6b06f #a18c4b #ece1b8 #c7bc93 #ada277 #8e8251 #9a9065
#> 33  qualitative #e7bee4 #e6b155 #c69e58 #dcbc79 #a49c8b #f6cb68 #978759 #f5db8a
#> 34    diverging #9d7b51 #bf9259 #d9ad71 #766752 #bba481 #9f8f75 #d9c096 #847b68
#> 35  qualitative #5c5055 #88736c #cb714e #9b8c86 #f2814f #7f5f50 #f4996b #a16c4e
#> 36     tricolor #4f494a #da7046 #a06348 #665750 #d2cecc #825f4d #be7547 #ab734c
#> 37  qualitative #ad717d #cba4a5 #766262 #5c4e4d #f0dcd7 #7a564c #bd9587 #e0c3b8
#> 38  qualitative #5A5C37 #209ACA #97AE3F #FFDF1D #4AB2CB #DA6641 #B89E92 #374228
#> 39     tricolor #5a6c5d #4c6254 #63746f #54706d #496764 #717a7b #465053 #ddf1f8
#> 40  qualitative #876374 #69505a #9a7a83 #ac999e #8d4b58 #bdb9b9 #d2cbc3 #dedbd6
#> 41  qualitative #7e6572 #a1969b #7b5252 #a67977 #a05c54 #bd7857 #cb9678 #bdb39a
#> 42    diverging #5f4a4b #988280 #765450 #88635d #a47063 #975c4d #b2877c #cf8d77
#> 43    diverging #d9c085 #cbb172 #bea564 #b29956 #a58d4b #917e49 #d8d0b5 #928861
#> 44    diverging #d6d2d1 #f4f1ef #cbc1af #a18952 #ada695 #81724d #c8aa5c #665f4b
#> 45    diverging #7b6c50 #6a5f4a #b9b3a7 #8a7b5d #998a6b #a89a7b #5a5549 #a9a391
#> 46   sequential #4c9de2 #5ea5e3 #4d94d4 #73aee4 #85baeb #9cc6ef #4987c4 #699dd4
#> 47     tricolor #a76a61 #b5ab94 #d7ce57 #d3d2be #d6d58f #b6b550 #b9bf74 #98a14f
#> 48  qualitative #a65258 #ca7f7a #94928f #b8b6b0 #605f5b #a6a49c #83827d #cac9c0
#> 49    diverging #777676 #525151 #786451 #908b85 #bc8b4d #957c58 #afa493 #f2d4a3
#> 50     tricolor #4e494d #615657 #746764 #d48974 #867c77 #96918e #c5bab3 #aaa4a0
#> 51   sequential #5e4f47 #6d5f57 #775947 #aa734e #817166 #806755 #8d6648 #b98c69
#> 52    diverging #efc66d #bfa55f #969256 #c0bf7f #efefb2 #d9db9a #718353 #97b17d
#> 53  qualitative #b97b75 #dd9489 #997551 #736550 #c6a260 #e4c58c #c1b68b #e5ddb7
#> 54  qualitative #d0cece #ef7d55 #886153 #b2aaa5 #ad8a6e #e3e1de #e5c580 #f3f3f1
#> 55    diverging #8e8a66 #78744f #c5c3a5 #a19f7b #b0af6a #9da053 #888d4c #667d4f
#>    color_9 color_10 color_11 color_12 color_13 color_14 color_15 color_16
#> 1  #c5c0a0  #d7d3b7  #eeeac7  #e5e192  #aaaaa9  #d7d95d  #bec07e  #a0a673
#> 2  #9a584d  #60524d  #7f7e53  #6a6c4d  #92974c  #a7b059  #8b916f  #aebf85
#> 3  #857e4d  #a8a16f  #767569  #f4f0ba  #c0bb72  #dcd9a3  #c7c58d  #b2af55
#> 4  #75d96e  #66b164  #56d067  #51995e  #497d55  #72dd96  #59d885  #a3edbd
#> 5  #ebe4e0  #af9b8e  #957a68  #8d664a  #c4b09e  #a98f72  #dec8aa  #e6b155
#> 6  #ac8368  #bdaca0  #af9885  #8e7560  #7c6852  #968874  #675c4b  #80845b
#> 7  #a29483  #b6a895  #998467  #877151  #eff3f6  #b7c1cd  #a0afc2  #4b5664
#> 8  #dc7d66  #cf6f4d  #93654a  #f3ddcf  #6a5e4a  #dfe8ea  #c2e0f0  #a9cae5
#> 9  #fc7946  #fc905b  #c5734a  #fd8b47  #e28852  #fba35d  #fd9d48  #fdb861
#> 10 #c8ad8e  #e4d5be  #978e7e  #7d786b  #e9d562  #eae9dd  #bcbdb6  #a1a49e
#> 11 #dde5f1  #c3cbd9  #75839e  #a7b9dc  #8695b3  #d5d9e1  #96a6c9  #606d8d
#> 12 #ad8e76  #918881  #9f7f66  #765f4d  #c8946b  #cdaf96  #bca088  #edbb90
#> 13 #daa97e  #b2a89f  #b7977a  #cb884a  #d09762  #998169  #e8a664  #9a948e
#> 14 #dab292  #b9a694  #6d6154  #d4cec4  #ede4d1  #6e844c  #7d9859  #8da969
#> 15 #b4b990  #c9cfa2  #d6eb4f  #bed551  #9da382  #c0d179  #a9c24f  #a7b96f
#> 16 #fab169  #fbc691  #fad8b1  #53504c  #fccf4e  #fddd6d  #fde88f  #49704e
#> 17 #967b69  #e5c3ac  #ba8b58  #63734e  #7fa755  #8ec659  #778a66  #a0db73
#> 18 #978c6e  #eef2e5  #727b6f  #a6d8a8  #8eb990  #4f5a50  #6f9c76  #c2e0c8
#> 19 #d5e2f8  #8a93a7  #646a78  #92a1c6  #c1cff2  #9eafdb  #acbeef  #8190b9
#> 20 #d8c5b2  #cab298  #c9a479  #eaa44f  #eeddc3  #e7c896  #f3f1e6  #96978a
#> 21 #fbf2af  #e1e0cb  #f4f4e1  #abab76  #abaf98  #8c986f  #c9cebd  #90b050
#> 22 #887566  #9a816c  #a59282  #605348  #4d4c4a  #c5b361  #5c5c5b  #a2b5d1
#> 23 #7b7c4e  #696a4d  #575948  #5a6b78  #5388cd  #d5e2f3  #6d9ad5  #9caec7
#> 24 #c59d76  #695a4b  #ae8c67  #a79d92  #af987e  #dfb98d  #f4c284  #998061
#> 25 #cdc8ab  #757053  #817c5c  #e8e5ce  #66655d  #d8d5bc  #bdba9d  #96947a
#> 26 #907350  #7f674a  #cea774  #c29a64  #b08d5c  #a18152  #e1b87a  #e9c995
#> 27 #b5a595  #d4a770  #d6b68e  #e0a151  #f7c67f  #f7b758  #fce05f  #bebcb2
#> 28 #647794  #7c8eaa  #71829d  #607ba6  #526e9b  #5a6c89  #757d8a  #4c6185
#> 29 #985f50  #e58a72  #ba6a51  #db7751  #f8a380  #f78e53  #928780  #80746a
#> 30 #a48d70  #e7caa5  #ba9f7a  #ceb38d  #81786a  #a4a093  #b6b7af  #818987
#> 31 #a98b54  #988e7b  #8d7b50  #73684d  #f4e6bc  #797972  #eff1e7  #545650
#> 32 #80764a  #e6d587  #6f6847  #4c5551  #8d9895  #637275  #edf2f6  #4a5a6d
#> 33 #bcb082  #6b664c  #a29b6f  #7c7852  #efeab9  #87856b  #c8cea9  #9fc86c
#> 34 #5a564a  #dfdfc3  #b3b99d  #f0f6e0  #c9d5ae  #bce37e  #a9d367  #cdef9a
#> 35 #b8a9a0  #cf8d64  #ac8163  #f7b783  #b9977b  #d3c7bd  #d4a87f  #daba9a
#> 36 #a17c62  #926e54  #bc8659  #d18b4a  #d39a5d  #e79f4c  #eab162  #dcc6a6
#> 37 #936f5b  #ae8a65  #ccc48e  #bfb75c  #9ba656  #a8b278  #87954d  #898f7a
#> 38    <NA>     <NA>     <NA>     <NA>     <NA>     <NA>     <NA>     <NA>
#> 39 #4a585e  #81c4eb  #c2e0f3  #586269  #a9d2f0  #5a7488  #68aee7  #5199e3
#> 40 #887f52  #a8b15d  #7c8a76  #bad1e0  #cde1f1  #a1c1db  #84afd2  #6586a3
#> 41 #bba151  #dbcea1  #d7bd56  #e0cc79  #948d51  #bbb872  #9ba376  #6d7654
#> 42 #c47c63  #b3674c  #e6926d  #cd7249  #e0875c  #de7d49  #ee894e  #f4955d
#> 43 #77704e  #b4af95  #9f9979  #bdc0ba  #51574d  #767d7f  #a8b0b5  #989ea2
#> 44 #9c9379  #867f6a  #716d5e  #efd369  #9eb5cd  #aecef3  #7ea5d5  #93bbed
#> 45 #6b675b  #787465  #868272  #959283  #494b4b  #565a5c  #62686c  #98a4ad
#> 46 #5c8fc5  #4b7eb4  #84aad3  #4c6e93  #99b9dc  #799abf  #5b7b9f  #8ea4bd
#> 47 #a0ad69  #7e944d  #e3e4e1  #afc994  #84a06c  #96af85  #60834d  #6a8d69
#> 48 #71716d  #9dbc4c  #ebedee  #d7dadc  #8b9aab  #7a889d  #91a6c6  #abb8cc
#> 49 #d2b27d  #ae966e  #edc681  #d1a556  #e7b861  #cfcac0  #b9b6af  #5590c0
#> 50 #d9d1ca  #6a747c  #5b656d  #79838c  #586c81  #8e9ba9  #495562  #70849a
#> 51 #917b69  #cea079  #937357  #a68c74  #a18163  #d09155  #b49a81  #c2a88e
#> 52 #b3d19c  #739f74  #4f7151  #b6e9bb  #93c7a7  #518c6c  #75b898  #d3f3e7
#> 53 #918a67  #a69f7a  #b7d3aa  #eaf7e6  #707a71  #9bb0a0  #515352  #7c9590
#> 54 #4a82b5  #618fbc  #6282a5  #7996b6  #84a7ce  #6a9ad0  #4b709a  #4e8ad0
#> 55 #54664b  #6e896a  #7c9983  #527669  #8aaca6  #5b8487  #6f989c  #dcedf5
#>    color_17 color_18 color_19 color_20 color_21 color_22 color_23 color_24
#> 1   #acb954  #909a52  #748050  #818d6f  #65666c  #d5d6de  #888ac4  #bebdc3
#> 2   #6b7d9e  #596783  #9aa8c8  #d8dce7  #858fad  #505469  #bfc0d6  #797384
#> 3   #9a984f  #f0ee92  #dcdb78  #f1f1e9  #cace5a  #515552  #b3c7ec  #8799c3
#> 4   #4abe73  #8adfaa  #66bc90  #549f8a  #80b2a5  #465b57  #728c8f  #5e707e
#> 5   #9b7f4c  #c7ab74  #c9a04d  #b0914f  #747a98  #9596a0  #5e6186  #4d4e73
#> 6   #6f7748  #90a8c1  #7691b5  #b1b4bc  #565863  #6a66a9  #514b9a  #8b7ebd
#> 7   #727d8c  #7f99bd  #5a6677  #cad1db  #8a9099  #dde1e7  #9da0a5  #707176
#> 8   #7da7cf  #c6d0dd  #9bb2d3  #647cb0  #909ac1  #7d87b7  #b8bcd2  #9d83ad
#> 9   #fdb049  #fed57a  #fdc64a  #fed85f  #fcde47  #fdec52  #fdf56c  #fdfab9
#> 10  #505351  #bad8dc  #758a97  #7ba7c3  #576874  #69b3eb  #52769c  #5589c0
#> 11  #babec8  #acb0bb  #9da1ac  #8d9099  #53586a  #f1f2f7  #696971  #7c7b83
#> 12  #f3a561  #dfa776  #d8c1a9  #544f49  #f8be75  #ecd3b3  #bcbab7  #fcec7c
#> 13  #e4994b  #f5cb9f  #76614b  #f8ba76  #f6ae53  #81725e  #bbc1c2  #a2afbd
#> 14  #94be85  #678092  #849db2  #586370  #b1c0d4  #c6d2e6  #a9adb8  #898993
#> 15  #94ae4f  #8fa06a  #809a4e  #6c864f  #7a8c66  #5a774c  #858e82  #4d604c
#> 16  #d9d8e6  #cbc7d8  #bcb7cc  #5e5969  #afa7c0  #716a7f  #a397b1  #867a91
#> 17  #b4eb8d  #9abc82  #67914e  #85a479  #c0ebb3  #a5c9a3  #9ba2a1  #bec2c6
#> 18  #547762  #849692  #c0edf2  #a6b5b7  #7da9b5  #58818d  #95cbec  #66a6ed
#> 19  #525c79  #7480a4  #b7bed7  #616b91  #4c5065  #abaec2  #a1a0ae  #87858f
#> 20  #bac1b1  #a9afa3  #c9d0c8  #d3e1e1  #adb5c8  #979aa9  #828492  #73737b
#> 21  #7e994d  #5e6c4b  #7a866d  #69844d  #697260  #919c89  #4e594a  #b3bdb5
#> 22  #b4c1d6  #92a5c4  #8a93a5  #7d8596  #73767e  #aaaebb  #cacdd9  #69696a
#> 23  #89acdd  #a9bedc  #becee5  #4c77b7  #818995  #6989b9  #4d6891  #849bbe
#> 24  #968a7b  #d5ad78  #f8d098  #827663  #7f6b4b  #bf9e60  #fce4af  #ab8f50
#> 25  #8a886b  #f7f6e5  #a0a08a  #acad9d  #75776c  #babcb4  #7e8882  #c4d3df
#> 26  #ddd0ae  #e0dfce  #c9cbbe  #b7bbb9  #a3a7ab  #637d9b  #4f698d  #7c889b
#> 27  #fcef9b  #9caaaf  #7891a9  #647294  #969698  #4f4f7a  #4e4d4f  #5d5b5e
#> 28  #586479  #485874  #828b9b  #687181  #475061  #555b68  #47484d  #66676e
#> 29  #fcba56  #fdd975  #6d6e52  #87aac8  #5b8abf  #bcc3ce  #8f96a2  #737c8c
#> 30  #c8cfcf  #88a2af  #9fbccb  #d3e4ef  #a6d0f0  #6485a2  #78afe3  #548bce
#> 31  #556974  #d2d8de  #7a8999  #5884ba  #b1cbf0  #bbc2ce  #92b6f8  #7298f2
#> 32  #4f6584  #607491  #c4cedd  #70829e  #a3afc2  #7f91ae  #91a0b9  #b3bdd0
#> 33  #83b053  #6d9550  #a7ca94  #7b9f6a  #505e4a  #8db57d  #597d4b  #647e6a
#> 34  #9daa8a  #8aa366  #93bf54  #9bba6f  #80a74d  #6f8a4d  #daf6b7  #a9c984
#> 35  #f3cca3  #f9ac54  #f5e2c0  #f4efe2  #fbd363  #fdea8a  #bfd0f2  #879cd0
#> 36  #f9b44e  #fac668  #f6ce83  #f5e3bc  #fadf9d  #f2ecda  #74819e  #b4b6c0
#> 37  #6e7d4c  #577466  #5a828d  #7c98a2  #96a7af  #5a9dd0  #80aacf  #a1b5c9
#> 38     <NA>     <NA>     <NA>     <NA>     <NA>     <NA>     <NA>     <NA>
#> 39  #4d657f  #5182b8  #485b70  #6a809a  #97bdeb  #83aae3  #7e96ba  #898895
#> 40  #55728c  #7c99b5  #99a8bd  #4f5b6d  #747a8a  #8b8f9c  #f0f0f1  #676773
#> 41  #7a8a72  #535553  #dadbe0  #7f86b4  #545878  #646996  #9ea1c0  #bbbcc9
#> 42  #f7a06c  #f9b179  #fac28b  #6a7785  #4d5766  #5c6573  #4b4750  #817e84
#> 43  #888e92  #536470  #d4dbe0  #89b0cd  #5d768a  #7392ab  #adc5d9  #c1cfdc
#> 44  #8aa1be  #6e90bb  #4e5b6f  #b8c3d4  #c7d4e9  #606e85  #778196  #e0e3ec
#> 45  #808589  #71767a  #89959f  #a6b2bc  #bcc2c7  #70859a  #546e89  #cdd1d6
#> 46  #4e637b  #9eb3cb  #698ab0  #607287  #7f93ab  #70839a  #475567  #b1c5df
#> 47  #b6b9b6  #4d6955  #788c97  #93a4b0  #577187  #568cc8  #6fa2dc  #8fb6e3
#> 48  #a1a9b6  #7a93bf  #6580b0  #6c778b  #c0c8d7  #59657e  #4c566c  #586d9d
#> 49  #527492  #5a636b  #7aa1c5  #748a9f  #979fa7  #9db3ca  #d4d9df  #b7c5d6
#> 50  #818f9f  #4a5e75  #66788d  #9aa6b5  #a9b2be  #b8bfcb  #c7cdd8  #dbdde5
#> 51  #d0b79d  #948677  #aea293  #a09586  #d6cab9  #dfd6c9  #bcb1a1  #cabfaf
#> 52  #7ddfbc  #519f90  #56bfab  #50d8cc  #93ecef  #5ee3f3  #79c4cf  #4fbfed
#> 53  #bbe6e0  #97cbc8  #79b0b5  #89e5ee  #70c5da  #5d93a6  #59a5c8  #54788e
#> 54  #93a4b9  #a2b3c9  #545962  #54617b  #4b4c50  #94959b  #bebebf  #818188
#> 55  #6fa0b9  #558ba8  #8ab5cd  #95c7eb  #74b7e9  #b5d6ef  #54a4e1  #6da4d2
#>    color_25
#> 1   #7d7882
#> 2   #4d484e
#> 3   #5e6d93
#> 4   #555272
#> 5   #aaa8ae
#> 6   #a395ce
#> 7   #4b4b4c
#> 8   #8b6c93
#> 9   #fdf88b
#> 10  #5299e5
#> 11  #4d4b51
#> 12  #cfd0cc
#> 13  #8a9cb0
#> 14  #9d9ba1
#> 15  #667669
#> 16  #9987a2
#> 17  #d6dbe3
#> 18  #5389c5
#> 19  #7c7582
#> 20  #635f66
#> 21  #9baaaf
#> 22  #dcdce2
#> 23  #5a759f
#> 24  #95804a
#> 25  #aebdcb
#> 26  #9295a0
#> 27  #674b6a
#> 28  #555456
#> 29  #5a6373
#> 30  #596370
#> 31  #9b9eaa
#> 32  #d7deec
#> 33  #7797b3
#> 34  #b9d797
#> 35  #5e6c9e
#> 36  #979baf
#> 37  #bdc7d5
#> 38     <NA>
#> 39  #b5adbb
#> 40  #4d4c55
#> 41  #847e93
#> 42  #726e75
#> 43  #e3e8ee
#> 44  #4d4d4d
#> 45  #dee1e8
#> 46  #c9d7ed
#> 47  #b4c6de
#> 48  #4d4d50
#> 49  #eaedf1
#> 50  #eff0f5
#> 51  #8e834e
#> 52  #4ca2c7
#> 53  #4f6172
#> 54  #6e6d74
#> 55  #528ec6

# What if our query returns NO options?
palette_find(name = "no such name")
#> No palette met the conditions. Run `palette_find()` to see all available palette options
#> [1] photographer      palette_full_name lter_site         palette_name     
#> [5] palette_type     
#> <0 rows> (or 0-length row.names)

# What if our query returns MULTIPLE options?
palette_find(site = "sbc")
#> Multiple options returned as a dataframe. Consider identifying a specific palette's name and re-running function
#>   photographer palette_full_name lter_site palette_name palette_type color_1
#> 1     SBC LTER   SBC-kelp forest       SBC  kelp forest    diverging #efc66d
#> 2   Kyle Emery      SBC-kelp tag       SBC     kelp tag  qualitative #b97b75
#>   color_2 color_3 color_4 color_5 color_6 color_7 color_8 color_9 color_10
#> 1 #bfa55f #969256 #c0bf7f #efefb2 #d9db9a #718353 #97b17d #b3d19c  #739f74
#> 2 #dd9489 #997551 #736550 #c6a260 #e4c58c #c1b68b #e5ddb7 #918a67  #a69f7a
#>   color_11 color_12 color_13 color_14 color_15 color_16 color_17 color_18
#> 1  #4f7151  #b6e9bb  #93c7a7  #518c6c  #75b898  #d3f3e7  #7ddfbc  #519f90
#> 2  #b7d3aa  #eaf7e6  #707a71  #9bb0a0  #515352  #7c9590  #bbe6e0  #97cbc8
#>   color_19 color_20 color_21 color_22 color_23 color_24 color_25
#> 1  #56bfab  #50d8cc  #93ecef  #5ee3f3  #79c4cf  #4fbfed  #4ca2c7
#> 2  #79b0b5  #89e5ee  #70c5da  #5d93a6  #59a5c8  #54788e  #4f6172

# What if our query returns JUST ONE option? (this is desirable)
palette_find(name = "salamander")
#> Exactly one palette identified. Output cropped to only HEX codes for ease of plotting
#>  [1] "#6c616a" "#957b8e" "#bfa3b7" "#ab8da1" "#8a686f" "#b17b82" "#7b5350"
#>  [8] "#af6a5f" "#9a584d" "#60524d" "#7f7e53" "#6a6c4d" "#92974c" "#a7b059"
#> [15] "#8b916f" "#aebf85" "#6b7d9e" "#596783" "#9aa8c8" "#d8dce7" "#858fad"
#> [22] "#505469" "#bfc0d6" "#797384" "#4d484e"